+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - PHP 8.0 support
+- Migration trigger on module activation
 
 ## [3.0.0] - 2023-06-08